Make sure we can checkout a credential based repo.
This uses the rightscale-test-account on github.
The down side here is that this pub key is known,
now, so folks could in theory abuse this.  But it
doesn't have any information we care about.

This test is only run if ANNOY_GITHUB is turned on
as an attempt at mollifying this silliness.